Put the new import_model() in its own Python file
